[ArticleListPage] extract components to under components/
